Backend/mcp-server/requirements.txt
# CarFin-MCP Server Dependencies - 실제 AI 모델 포함

# 웹 프레임워크 (필수)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0

# 데이터베이스 (필수)
asyncpg==0.29.0

# 실제 AI/ML 모델들
torch==2.1.1
transformers==4.35.2
tokenizers==0.15.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

# KoBERT 한국어 감정분석
kobert-transformers==0.5.1
sentencepiece==0.1.99

# LSTM + XGBoost 감가상각 예측
tensorflow==2.15.0
xgboost==2.0.2

# 추천시스템
surprise==1.1.3
implicit==0.7.2

# Google Cloud AI/ML
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0

# HTTP 클라이언트 (필수)
httpx==0.25.2

# 유틸리티 (필수)
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1